+++ server/src/main/java/org/opensearch/index/mapper/DateFieldMapper.java
@@ -53,6 +53,7 @@ import org.opensearch.common.time.DateFormatters;
 import org.opensearch.common.time.DateMathParser;
 import org.opensearch.common.time.DateUtils;
 import org.opensearch.common.unit.TimeValue;
+import org.opensearch.common.util.FeatureFlags;
 import org.opensearch.common.util.LocaleUtils;
 import org.opensearch.index.fielddata.IndexFieldData;
 import org.opensearch.index.fielddata.IndexNumericFieldData.NumericType;
@@ -92,7 +93,21 @@ public final class DateFieldMapper extends ParametrizedFieldMapper {
 
     public static final String CONTENT_TYPE = "date";
     public static final String DATE_NANOS_CONTENT_TYPE = "date_nanos";
-    public static final DateFormatter DEFAULT_DATE_TIME_FORMATTER = DateFormatter.forPattern("strict_date_optional_time||epoch_millis");
+    @Deprecated
+    public static final DateFormatter LEGACY_DEFAULT_DATE_TIME_FORMATTER = DateFormatter.forPattern(
+        // TODO remove in 3.0 after backporting
+        "strict_date_optional_time||epoch_millis"
+    );
+    public static final DateFormatter DEFAULT_DATE_TIME_FORMATTER = DateFormatter.forPattern(
+        "strict_date_time_no_millis||strict_date_optional_time||epoch_millis",
+        "strict_date_optional_time"
+    );
+
+    public static DateFormatter getDefaultDateTimeFormatter() {
+        return FeatureFlags.isEnabled(FeatureFlags.DATETIME_FORMATTER_CACHING_SETTING)
+            ? DEFAULT_DATE_TIME_FORMATTER
+            : LEGACY_DEFAULT_DATE_TIME_FORMATTER;
+    }
 
     /**
      * Resolution of the date time
@@ -224,8 +239,14 @@ public final class DateFieldMapper extends ParametrizedFieldMapper {
             "format",
             false,
             m -> toType(m).format,
-            DEFAULT_DATE_TIME_FORMATTER.pattern()
+            getDefaultDateTimeFormatter().pattern()
         );
+        private final Parameter<String> printFormat = Parameter.stringParam(
+            "print_format",
+            false,
+            m -> toType(m).printFormat,
+            getDefaultDateTimeFormatter().printPattern()
+        ).acceptsNull();
         private final Parameter<Locale> locale = new Parameter<>(
             "locale",
             false,
@@ -254,6 +275,7 @@ public final class DateFieldMapper extends ParametrizedFieldMapper {
             this.ignoreMalformed = Parameter.boolParam("ignore_malformed", true, m -> toType(m).ignoreMalformed, ignoreMalformedByDefault);
             if (dateFormatter != null) {
                 this.format.setValue(dateFormatter.pattern());
+                this.printFormat.setValue(dateFormatter.printPattern());
                 this.locale.setValue(dateFormatter.locale());
             }
         }
@@ -262,9 +284,13 @@ public final class DateFieldMapper extends ParametrizedFieldMapper {
             try {
                 if (Joda.isJodaPattern(indexCreatedVersion, format.getValue())) {
                     return Joda.forPattern(format.getValue()).withLocale(locale.getValue());
-                } else {
-                    return DateFormatter.forPattern(format.getValue()).withLocale(locale.getValue());
                 }
+                if (format.isConfigured() && !printFormat.isConfigured()) {
+                    return DateFormatter.forPattern(format.getValue(), null, !format.isConfigured()).withLocale(locale.getValue());
+                }
+                return DateFormatter.forPattern(format.getValue(), printFormat.getValue(), !format.isConfigured())
+                    .withLocale(locale.getValue());
+
             } catch (IllegalArgumentException e) {
                 throw new IllegalArgumentException("Error parsing [format] on field [" + name() + "]: " + e.getMessage(), e);
             }
@@ -272,7 +298,7 @@ public final class DateFieldMapper extends ParametrizedFieldMapper {
 
         @Override
         protected List<Parameter<?>> getParameters() {
-            return Arrays.asList(index, docValues, store, format, locale, nullValue, ignoreMalformed, boost, meta);
+            return Arrays.asList(index, docValues, store, format, printFormat, locale, nullValue, ignoreMalformed, boost, meta);
         }
 
         private Long parseNullValue(DateFieldType fieldType) {
@@ -351,7 +377,7 @@ public final class DateFieldMapper extends ParametrizedFieldMapper {
         }
 
         public DateFieldType(String name) {
-            this(name, true, false, true, DEFAULT_DATE_TIME_FORMATTER, Resolution.MILLISECONDS, null, Collections.emptyMap());
+            this(name, true, false, true, getDefaultDateTimeFormatter(), Resolution.MILLISECONDS, null, Collections.emptyMap());
         }
 
         public DateFieldType(String name, DateFormatter dateFormatter) {
@@ -359,7 +385,7 @@ public final class DateFieldMapper extends ParametrizedFieldMapper {
         }
 
         public DateFieldType(String name, Resolution resolution) {
-            this(name, true, false, true, DEFAULT_DATE_TIME_FORMATTER, resolution, null, Collections.emptyMap());
+            this(name, true, false, true, getDefaultDateTimeFormatter(), resolution, null, Collections.emptyMap());
         }
 
         public DateFieldType(String name, Resolution resolution, DateFormatter dateFormatter) {
@@ -615,6 +641,7 @@ public final class DateFieldMapper extends ParametrizedFieldMapper {
     private final boolean hasDocValues;
     private final Locale locale;
     private final String format;
+    private final String printFormat;
     private final boolean ignoreMalformed;
     private final Long nullValue;
     private final String nullValueAsString;
@@ -638,6 +665,7 @@ public final class DateFieldMapper extends ParametrizedFieldMapper {
         this.hasDocValues = builder.docValues.getValue();
         this.locale = builder.locale.getValue();
         this.format = builder.format.getValue();
+        this.printFormat = builder.printFormat.getValue();
         this.ignoreMalformed = builder.ignoreMalformed.getValue();
         this.nullValueAsString = builder.nullValue.getValue();
         this.nullValue = nullValue;